+++ app/mainwindow.cpp
@@ -2,6 +2,7 @@
 #include "ui_mainwindow.h"
 #include "config/storage.h"
 #include "waitingspinnerwidget.h"
+#include "shortcutsdialog.h"
 
 #include <QDebug>
 #include <QApplication>
@@ -346,32 +347,10 @@ void MainWindow::setupShortcuts() {
   connect(shortcuts, &Shortcuts::openSortMenu, ui->sortButton, &QToolButton::click);
 
   connect(main_menu, &MainMenu::openShortcuts, [=]() {
-    QDialog dlg(this);
-    QVBoxLayout layout(&dlg);
-    QTableWidget tbl(&dlg);
-    auto data = shortcuts->toStringHash();
-    tbl.setRowCount(data.size());
-    tbl.setColumnCount(2);
-    tbl.verticalHeader()->hide();
-    tbl.horizontalHeader()->hide();
-    tbl.horizontalHeader()->setSectionResizeMode(0, QHeaderView::Stretch);
-    tbl.horizontalHeader()->setSectionResizeMode(1, QHeaderView::Stretch);
-    QHashIterator<QString, QString> it(data);
-    int row = 0;
-    while (it.hasNext()) {
-      it.next();
-      auto item1 = new QTableWidgetItem(it.key());
-      auto item2 = new QTableWidgetItem(it.value());
-      item1->setFlags(Qt::ItemIsEnabled);
-      item2->setFlags(Qt::ItemIsEnabled);
-      tbl.setItem(row, 0, item1);
-      tbl.setItem(row, 1, item2);
-      row++;
-    }
-    layout.addWidget(&tbl);
-    dlg.setLayout(&layout);
-    dlg.resize(dlg.sizeHint());
-    dlg.exec();
+    auto dlg = new ShortcutsDialog(shortcuts, this);
+    dlg->setModal(false);
+    connect(dlg, &ShortcutsDialog::finished, dlg, &ShortcutsDialog::deleteLater);
+    dlg->show();
   });
 }